+++ src/Twilio/Rest/Api/V2010/Account/RecordingList.php
@@ -98,6 +98,7 @@ class RecordingList extends ListResource {
             'DateCreated>' => Serialize::iso8601DateTime($options['dateCreatedAfter']),
             'CallSid' => $options['callSid'],
             'ConferenceSid' => $options['conferenceSid'],
+            'IncludeSoftDeleted' => Serialize::booleanToString($options['includeSoftDeleted']),
             'PageToken' => $pageToken,
             'Page' => $pageNumber,
             'PageSize' => $pageSize,