+++ src/Twilio/Rest/Chat/V2/Service/User/UserChannelInstance.php
@@ -97,11 +97,12 @@ class UserChannelInstance extends InstanceResource {
     /**
      * Delete the UserChannelInstance
      *
+     * @param array|Options $options Optional Arguments
      * @return bool True if delete succeeds, false otherwise
      * @throws TwilioException When an HTTP error occurs.
      */
-    public function delete(): bool {
-        return $this->proxy()->delete();
+    public function delete(array $options = []): bool {
+        return $this->proxy()->delete($options);
     }
 
     /**