+++ tests/Twilio/Integration/Numbers/V2/RegulatoryCompliance/SupportingDocumentTest.php
@@ -45,6 +45,7 @@ class SupportingDocumentTest extends HolodeckTestCase {
                 "friendly_name": "friendly_name",
                 "mime_type": "mime_type",
                 "status": "draft",
+                "failure_reason": null,
                 "type": "type",
                 "attributes": {
                     "first_name": "foo",
@@ -115,6 +116,7 @@ class SupportingDocumentTest extends HolodeckTestCase {
                         "friendly_name": "friendly_name",
                         "mime_type": "mime_type",
                         "status": "draft",
+                        "failure_reason": null,
                         "type": "type",
                         "attributes": {
                             "first_name": "foo",
@@ -144,6 +146,48 @@ class SupportingDocumentTest extends HolodeckTestCase {
         $this->assertGreaterThan(0, \count($actual));
     }
 
+    public function testReadRejectedDocumentResponse(): void {
+        $this->holodeck->mock(new Response(
+            200,
+            '
+            {
+                "results": [
+                    {
+                        "sid": "RDaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "friendly_name": "friendly_name",
+                        "mime_type": "mime_type",
+                        "status": "twilio-rejected",
+                        "failure_reason": "Some failure reason.",
+                        "type": "type",
+                        "attributes": {
+                            "first_name": "foo",
+                            "last_name": "bar"
+                        },
+                        "date_created": "2019-07-31T02:11:52Z",
+                        "date_updated": "2019-07-31T02:11:52Z",
+                        "url": "https://numbers.twilio.com/v2/RegulatoryCompliance/SupportingDocuments/RDaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                    }
+                ],
+                "meta": {
+                    "page": 0,
+                    "page_size": 50,
+                    "first_page_url": "https://numbers.twilio.com/v2/RegulatoryCompliance/SupportingDocuments?PageSize=50&Page=0",
+                    "previous_page_url": null,
+                    "url": "https://numbers.twilio.com/v2/RegulatoryCompliance/SupportingDocuments?PageSize=50&Page=0",
+                    "next_page_url": null,
+                    "key": "results"
+                }
+            }
+            '
+        ));
+
+        $actual = $this->twilio->numbers->v2->regulatoryCompliance
+                                            ->supportingDocuments->read();
+
+        $this->assertNotNull($actual);
+    }
+
     public function testFetchRequest(): void {
         $this->holodeck->mock(new Response(500, ''));
 
@@ -169,6 +213,7 @@ class SupportingDocumentTest extends HolodeckTestCase {
                 "friendly_name": "friendly_name",
                 "mime_type": "mime_type",
                 "status": "draft",
+                "failure_reason": null,
                 "type": "type",
                 "attributes": {
                     "first_name": "foo",
@@ -212,6 +257,7 @@ class SupportingDocumentTest extends HolodeckTestCase {
                 "friendly_name": "friendly_name",
                 "mime_type": "mime_type",
                 "status": "draft",
+                "failure_reason": null,
                 "type": "type",
                 "attributes": {
                     "first_name": "foo",