+++ tests/Twilio/Unit/Rest/ClientTest.php
@@ -241,4 +241,20 @@ class ClientTest extends UnitTest {
         $this->assertStringNotContainsString('Authorization-header', stream_get_contents($capturedLogging));
     }
 
+    public function testDefaultUserAgent(): void {
+        $client = new Client('username', 'password');
+        $client->request('GET', 'https://api.twilio.com');
+        $userAgent = $client->getHttpClient()->{'lastRequest'}[CURLOPT_HTTPHEADER][0];
+        $this->assertRegExp('/^User-Agent: twilio-php\/[0-9.]+\s\(\w+\s\w+\)\sPHP\/[^\s]+$/',$userAgent);
+    }
+
+    public function testUserAgentExtensionsWhenSet(): void {
+        $expectedExtensions = ['twilio-run/2.0.0-test', 'flex-plugin/3.4.0'];
+        $client = new Client('username', 'password', null,null,null,null,$expectedExtensions);
+        $client->request('GET', 'https://api.twilio.com');
+        $userAgent = $client->getHttpClient()->{'lastRequest'}[CURLOPT_HTTPHEADER][0];
+        $userAgentExtensions = array_slice(explode(" ",$userAgent),-count($expectedExtensions));
+        $this->assertEquals($userAgentExtensions,$expectedExtensions);
+    }
+
 }